% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradDescAutocontenido.R
\docType{package}
\name{GradDescAutocontenido-package}
\alias{GradDescAutocontenido-package}
\title{GradDescAutocontenido}
\description{
GradDesc-package: paquete wrapper de la biblioteca Purple
}
\details{
La documentacion de la clase \code{GradientDescentSumaCuadradosClase} está en: \code{\link{GradientDescentSumaCuadradosClase}}
}
\author{
Gonzalo Mateo, Paulino Tardáguila
}

