% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradientDescentSumaCuadrados}
\alias{GradientDescentSumaCuadrados}
\title{GradientDescentSumaCuadrados}
\usage{
GradientDescentSumaCuadrados(A, b)
}
\arguments{
\item{A}{Matriz con regresores}

\item{b}{Vector objetivo}
}
\description{
GradientDescentSumaCuadrados
}

